Improved removal of old episodes, misc clean-ups, Unicode
Complete rewrite of the deleting old episodes.  Much more robust and quicker.
"recent" is configurable for the number of episodes to keep.  No interface yet but just a single variable to change.
Had to move all the way to Unicode in order to get extended characters to work on all platforms.  
Cleaned up status printing to screen

No known issues